Refactor session handshake layering
