+++ crates/transport/src/session/handshake/mod.rs
@@ -1,0 +1,9 @@
+mod negotiate;
+mod session;
+
+pub use negotiate::{
+    negotiate_session, negotiate_session_from_stream, negotiate_session_parts,
+    negotiate_session_parts_from_stream, negotiate_session_parts_with_sniffer,
+    negotiate_session_with_sniffer,
+};
+pub use session::SessionHandshake;
